requirements.txt
streamlit==1.40.2
langchain-core>=0.1.31,<0.2.0
langchain-community==0.0.28
langchain==0.0.28
tinydb==4.7.0
requests==2.28.1
streamlit-extras==0.3.6
deep_translator==1.11.4
streamlit-float==0.3.2
streamlit-chat==0.1.1
Pillow==10.4.0
st-pages==0.3.4
googletrans==4.0.0-rc1
streamlit-option-menu==0.4.0
passlib==1.7.4
langchain-text-splitters==0.0.1
flowise
redis==5.0.1
pymongo==4.6.1
gunicorn==21.2.0
uvicorn==0.27.0
fastapi==0.109.0
python-multipart==0.0.6
langdetect==1.0.9